@@ -4,7 +4,7 @@ data class PhotoEntity(
     val height: Int,
     val id: Int,
     val photographer: String,
-    val photographerId: Int,
+    val photographerId: Long,
     val photographerUrl: String,
     val src: PhotoSrcEntity,
     val url: String,
