@@ -1,22 +1,27 @@
 package com.greylabsdev.pexwalls.presentation.base
 
 import android.os.Bundle
+import android.view.LayoutInflater
 import androidx.annotation.IdRes
 import androidx.annotation.LayoutRes
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.Navigation
+import androidx.viewbinding.ViewBinding
 import java.io.Serializable
+import androidx.navigation.findNavController
 
-abstract class BaseActivity(
-    @LayoutRes private val layoutResId: Int,
+abstract class BaseActivity<VB: ViewBinding>(
+    private val bindingFactory: (inflater: LayoutInflater) -> VB,
     @IdRes private val navigationHostId: Int? = null
 ) : AppCompatActivity() {
 
+    protected lateinit var binding: VB
     protected abstract val viewModel: BaseViewModel?
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(layoutResId)
+        binding = bindingFactory(layoutInflater)
+        setContentView(binding.root)
 
         initViews()
         initListeners()
@@ -38,9 +43,9 @@ abstract class BaseActivity(
             navigationArgs?.let { args ->
                 val bundle = Bundle()
                 args.forEach { bundle.putSerializable(it.first, it.second) }
-                Navigation.findNavController(this, hostid).navigate(destinationId, bundle)
+                this.findNavController(hostid).navigate(destinationId, bundle)
             } ?: run {
-                Navigation.findNavController(this, hostid).navigate(destinationId)
+                this.findNavController(hostid).navigate(destinationId)
             }
         }
     }
