@@ -10,32 +10,33 @@ import android.view.View
 import android.view.ViewGroup
 import android.view.WindowManager
 import androidx.annotation.IdRes
-import androidx.annotation.LayoutRes
 import androidx.core.content.ContextCompat.getColor
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
-import androidx.navigation.Navigation
-import com.google.android.material.appbar.AppBarLayout
+import androidx.navigation.findNavController
+import androidx.viewbinding.ViewBinding
 import com.greylabsdev.pexwalls.R
+import com.greylabsdev.pexwalls.databinding.LayoutToolbarBinding
 import com.greylabsdev.pexwalls.presentation.view.PlaceholderView
 import java.io.Serializable
-import kotlinx.android.synthetic.main.layout_toolbar.*
-import kotlinx.android.synthetic.main.layout_toolbar.view.*
 
-abstract class BaseFragment(
-    @LayoutRes private val layoutResId: Int,
+abstract class BaseFragment<VB : ViewBinding>(
+    private val bindingFactory: (LayoutInflater, ViewGroup?, Boolean) -> VB,
     private val hasToolbarBackButton: Boolean = false,
     private val transparentStatusBar: Boolean = false,
     private val hideNavigation: Boolean = false
 ) : Fragment() {
 
+    private var _binding: VB? = null
+    protected val binding: VB?
+        get() = _binding
+
     protected abstract val viewModel: BaseViewModel?
     protected abstract val toolbarTitle: String?
     protected abstract val placeholderView: PlaceholderView?
     protected abstract val contentView: View?
-
-    private val toolbarView: AppBarLayout? by lazy { toolbar_container }
+    protected var toolbarView: LayoutToolbarBinding? = null
 
     private var onPermissionGrantedAction: (() -> Unit)? = null
 
@@ -45,7 +46,9 @@ abstract class BaseFragment(
         savedInstanceState: Bundle?
     ): View? {
         setupSystemBars()
-        return inflater.inflate(layoutResId, container, false)
+        _binding = bindingFactory(layoutInflater, container, false)
+        val view = binding?.root
+        return view
     }
 
     override fun onStart() {
@@ -67,18 +70,22 @@ abstract class BaseFragment(
                     placeholderView?.setState(PlaceholderView.PlaceholderState.GONE)
                     contentView?.isVisible = true
                 }
+
                 is ProgressState.LOADING -> {
                     placeholderView?.setState(PlaceholderView.PlaceholderState.LOADING)
                     contentView?.isVisible = false
                 }
+
                 is ProgressState.ERROR -> {
                     placeholderView?.setState(PlaceholderView.PlaceholderState.ERROR)
                     contentView?.isVisible = false
                 }
+
                 is ProgressState.INITIAL -> {
                     placeholderView?.setState(PlaceholderView.PlaceholderState.INITIAL)
                     contentView?.isVisible = false
                 }
+
                 is ProgressState.EMPTY -> {
                     placeholderView?.setState(PlaceholderView.PlaceholderState.EMPTY)
                     contentView?.isVisible = false
@@ -90,7 +97,7 @@ abstract class BaseFragment(
     protected open fun doInitialCalls() {}
 
     protected fun navigateBack() {
-        Navigation.findNavController(requireView()).popBackStack()
+        requireView().findNavController().popBackStack()
     }
 
     protected fun navigateTo(
@@ -100,19 +107,15 @@ abstract class BaseFragment(
         navigationArgs?.let { args ->
             val bundle = Bundle()
             args.forEach { bundle.putSerializable(it.first, it.second) }
-            Navigation.findNavController(requireView()).navigate(destinationId, bundle)
+            requireView().findNavController().navigate(destinationId, bundle)
         } ?: run {
-            Navigation.findNavController(requireView()).navigate(destinationId)
+            requireView().findNavController().navigate(destinationId)
         }
     }
 
     protected fun requestStoragePermissionWithAction(permissionNeededAction: () -> Unit) {
         onPermissionGrantedAction = permissionNeededAction
-        if (Build.VERSION.SDK_INT >= 23) {
-            requestPermissions(arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE), PERMISSION_CODE)
-        } else {
-            onPermissionGrantedAction?.invoke()
-        }
+        requestPermissions(arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE), PERMISSION_CODE)
     }
 
     override fun onRequestPermissionsResult(
@@ -121,7 +124,8 @@ abstract class BaseFragment(
         grantResults: IntArray
     ) {
         if (requestCode == PERMISSION_CODE &&
-            grantResults.first() == PackageManager.PERMISSION_GRANTED) {
+            grantResults.first() == PackageManager.PERMISSION_GRANTED
+        ) {
             onPermissionGrantedAction?.invoke()
             onPermissionGrantedAction = null
         }
@@ -129,17 +133,17 @@ abstract class BaseFragment(
 
     private fun initToolbar() {
         toolbarView?.let {
-            toolbar_container.toolbar_title_tv.text = toolbarTitle ?: ""
-            toolbar_container.back_iv.isVisible = hasToolbarBackButton
+            it.toolbarTitleTv.text = toolbarTitle ?: ""
+            it.backIv.isVisible = hasToolbarBackButton
             if (hasToolbarBackButton) {
-                toolbar_container.back_iv.setOnClickListener { navigateBack() }
+                it.backIv.setOnClickListener { navigateBack() }
             }
         }
     }
 
     private fun setupSystemBars() {
         activity?.window?.statusBarColor = Color.WHITE
-        if (hideNavigation) (requireActivity() as BaseActivity).hideNavigation()
+        if (hideNavigation) (requireActivity() as BaseActivity<*>).hideNavigation()
         if (transparentStatusBar) {
             requireActivity().window.apply {
                 setFlags(
@@ -152,20 +156,21 @@ abstract class BaseFragment(
                 )
                 decorView.systemUiVisibility = 0
             }
-            (requireActivity() as BaseActivity).hideNavigation()
+            (requireActivity() as BaseActivity<*>).hideNavigation()
         } else {
             requireActivity().window.apply {
                 clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_STATUS)
                 clearFlags(WindowManager.LayoutParams.FLAG_TRANSLUCENT_NAVIGATION)
             }
-            if (hideNavigation.not())(requireActivity() as BaseActivity).showNavigation()
+            if (hideNavigation.not()) (requireActivity() as BaseActivity<*>).showNavigation()
         }
         if (Build.VERSION.SDK_INT > Build.VERSION_CODES.O) {
             requireActivity().window.apply {
                 navigationBarColor = getColor(requireContext(), R.color.colorBackground)
                 decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR
                 if (transparentStatusBar.not())
-                    decorView.systemUiVisibility = View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR or View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR
+                    decorView.systemUiVisibility =
+                        View.SYSTEM_UI_FLAG_LIGHT_NAVIGATION_BAR or View.SYSTEM_UI_FLAG_LIGHT_STATUS_BAR
             }
         }
     }
