@@ -1,15 +1,15 @@
 package com.greylabsdev.pexwalls.presentation.collection.photogrid
 
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.greylabsdev.pexwalls.databinding.ItemFooterBinding
 import com.greylabsdev.pexwalls.presentation.paging.PagingItem
 
 class PhotoGridFooterViewHolder(
-    view: View,
+    private val binding: ItemFooterBinding,
     val width: Int,
     val height: Int
-) : RecyclerView.ViewHolder(view) {
+) : RecyclerView.ViewHolder(binding.root) {
 
     fun bind(data: PagingItem.ItemData, useHalfOfHeight: Boolean = false) {
         itemView.layoutParams = ViewGroup.LayoutParams(
