@@ -1,20 +1,20 @@
 package com.greylabsdev.pexwalls.presentation.collection.photogrid
 
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.greylabsdev.pexwalls.databinding.ItemHeaderBinding
 import com.greylabsdev.pexwalls.presentation.paging.PagingItem
-import kotlinx.android.synthetic.main.item_header.view.*
 
 class PhotoGridHeaderViewHolder(
-    view: View,
+    private val binding: ItemHeaderBinding,
     width: Int,
     height: Int
-) : RecyclerView.ViewHolder(view) {
+) : RecyclerView.ViewHolder(binding.root) {
     init {
-        view.layoutParams = ViewGroup.LayoutParams(width, height)
+        binding.root.layoutParams = ViewGroup.LayoutParams(width, height)
     }
+
     fun bind(data: PagingItem.ItemData) {
-        itemView.number_tv.text = data.title
+        binding.numberTv.text = data.title
     }
 }
