@@ -1,19 +1,20 @@
 package com.greylabsdev.pexwalls.presentation.collection.photolist
 
-import android.view.View
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.greylabsdev.pexwalls.databinding.ItemHeaderBinding
 import com.greylabsdev.pexwalls.presentation.paging.PagingItem
-import kotlinx.android.synthetic.main.item_header.view.*
 
 class PhotoListHeaderViewHolder(
-    view: View,
+    private val binding: ItemHeaderBinding,
     height: Int
-) : RecyclerView.ViewHolder(view) {
+) : RecyclerView.ViewHolder(binding.root) {
     init {
-        view.layoutParams = ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, height)
+        binding.root.layoutParams =
+            ViewGroup.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, height)
     }
+
     fun bind(data: PagingItem.ItemData) {
-        itemView.number_tv.text = data.title
+        binding.numberTv.text = data.title
     }
 }
