@@ -5,10 +5,10 @@ import android.animation.ValueAnimator
 
 fun ValueAnimator.addOnEndAction(action: () -> Unit) {
     this.addListener(object : Animator.AnimatorListener {
-        override fun onAnimationEnd(p0: Animator?) { action.invoke() }
-        override fun onAnimationRepeat(p0: Animator?) {}
-        override fun onAnimationCancel(p0: Animator?) {}
-        override fun onAnimationStart(p0: Animator?) {}
+        override fun onAnimationEnd(p0: Animator) { action.invoke() }
+        override fun onAnimationRepeat(p0: Animator) {}
+        override fun onAnimationCancel(p0: Animator) {}
+        override fun onAnimationStart(p0: Animator) {}
     })
 }
 
