@@ -6,6 +6,7 @@ import android.view.View
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
 import com.greylabsdev.pexwalls.R
+import com.greylabsdev.pexwalls.databinding.FragmentFavoritesBinding
 import com.greylabsdev.pexwalls.presentation.base.BaseFragment
 import com.greylabsdev.pexwalls.presentation.collection.photogrid.PhotoGridPagingAdapter
 import com.greylabsdev.pexwalls.presentation.collection.photogrid.PhotoItemDecoration
@@ -16,17 +17,18 @@ import com.greylabsdev.pexwalls.presentation.ext.getScreenWidthInPixels
 import com.greylabsdev.pexwalls.presentation.model.PhotoModel
 import com.greylabsdev.pexwalls.presentation.screen.photo.PhotoFragment
 import com.greylabsdev.pexwalls.presentation.view.PlaceholderView
-import kotlinx.android.synthetic.main.fragment_favorites.*
-import org.koin.android.viewmodel.ext.android.viewModel
+import org.koin.androidx.viewmodel.ext.android.viewModel
 
-class FavoritesFragment : BaseFragment(
-    layoutResId = R.layout.fragment_favorites
+class FavoritesFragment : BaseFragment<FragmentFavoritesBinding>(
+    bindingFactory = FragmentFavoritesBinding::inflate
 ) {
 
     override val viewModel by viewModel<FavoritesViewModel>()
     override val toolbarTitle: String? by lazy { getString(R.string.favorites_toolbar_title) }
-    override val contentView: View? by lazy { photo_grid_rv }
-    override val placeholderView: PlaceholderView? by lazy { placeholder_view }
+    override val contentView: View?
+        get() = binding?.photoGridRv
+    override val placeholderView: PlaceholderView?
+        get() = binding?.placeholderView
 
     private val photoCardMargin by lazy { requireActivity().dpToPix(Consts.DEFAULT_MARGIN_DP) }
     private val photoCardWidth by lazy { requireActivity().getScreenWidthInPixels() / 2 }
@@ -41,17 +43,22 @@ class FavoritesFragment : BaseFragment(
     }
 
     override fun initViews() {
-        val staggeredGridLayoutManager = StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL)
-        photo_grid_rv.layoutManager = staggeredGridLayoutManager
-        photo_grid_rv.adapter = photoGridPagingAdapter
+        toolbarView = binding?.toolbar
+        val staggeredGridLayoutManager =
+            StaggeredGridLayoutManager(2, StaggeredGridLayoutManager.VERTICAL)
+        binding?.photoGridRv?.let { photoGrid ->
+            photoGrid.layoutManager = staggeredGridLayoutManager
+            photoGrid.adapter = photoGridPagingAdapter
 
-        if (photo_grid_rv.itemDecorationCount == 0) {
-            photo_grid_rv.addItemDecoration(
-                PhotoItemDecoration(
-                    photoCardMargin.toInt()
+            if (photoGrid.itemDecorationCount == 0) {
+                photoGrid.addItemDecoration(
+                    PhotoItemDecoration(
+                        photoCardMargin.toInt()
+                    )
                 )
-            )
+            }
         }
+
     }
 
     override fun initViewModelObserving() {
@@ -63,7 +70,7 @@ class FavoritesFragment : BaseFragment(
 
     override fun onPause() {
         super.onPause()
-        photo_grid_rv.layoutManager?.onSaveInstanceState()?.let { state ->
+        binding?.photoGridRv?.layoutManager?.onSaveInstanceState()?.let { state ->
             recyclerState = state
         }
     }
@@ -71,7 +78,7 @@ class FavoritesFragment : BaseFragment(
     override fun onResume() {
         super.onResume()
         recyclerState?.let { state ->
-            photo_grid_rv.layoutManager?.onRestoreInstanceState(state)
+            binding?.photoGridRv?.layoutManager?.onRestoreInstanceState(state)
         }
     }
 
