@@ -1,6 +1,6 @@
 package com.greylabsdev.pexwalls.presentation.screen.favorites
 
-import org.koin.android.viewmodel.dsl.viewModel
+import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 
 val favoritesModule = module {
