@@ -1,6 +1,7 @@
 package com.greylabsdev.pexwalls.presentation.screen.home.list
 
 import android.graphics.Outline
+import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.view.ViewOutlineProvider
@@ -8,9 +9,9 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.greylabsdev.pexwalls.R
+import com.greylabsdev.pexwalls.databinding.ItemCategoryThemeBinding
 import com.greylabsdev.pexwalls.presentation.ext.inflate
 import com.greylabsdev.pexwalls.presentation.model.CategoryModel
-import kotlinx.android.synthetic.main.item_category_theme.view.*
 
 class CategoryThemeAdapter(
     private val photoCardCornerRadius: Float,
@@ -25,7 +26,12 @@ class CategoryThemeAdapter(
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): CategoryThemeViewHolder {
         val view = parent.inflate(R.layout.item_category_theme, parent, false)
-        return CategoryThemeViewHolder(view, photoCardCornerRadius)
+        val itemBinding = ItemCategoryThemeBinding.inflate(
+            LayoutInflater.from(parent.context),
+            parent,
+            false
+        )
+        return CategoryThemeViewHolder(itemBinding, photoCardCornerRadius)
     }
 
     override fun getItemCount(): Int = categories.size
@@ -38,9 +44,9 @@ class CategoryThemeAdapter(
     }
 
     class CategoryThemeViewHolder(
-        view: View,
+        private val binding: ItemCategoryThemeBinding,
         private val photoCardCornerRadius: Float
-    ) : RecyclerView.ViewHolder(view) {
+    ) : RecyclerView.ViewHolder(binding.root) {
 
         private val outlineProvider by lazy {
             object : ViewOutlineProvider() {
@@ -49,14 +55,15 @@ class CategoryThemeAdapter(
                 }
             }
         }
+
         fun bind(categoryModel: CategoryModel) {
-            itemView.category_name_tv.text = categoryModel.category.name
+            binding.categoryNameTv.text = categoryModel.category.name
             Glide.with(itemView.context)
                 .load(categoryModel.categoryPhotoUrl)
                 .transform(CenterCrop())
-                .into(itemView.category_cover_iv)
-            itemView.category_cover_iv.outlineProvider = outlineProvider
-            itemView.category_cover_iv.clipToOutline = true
+                .into(binding.categoryCoverIv)
+            binding.categoryCoverIv.outlineProvider = outlineProvider
+            binding.categoryCoverIv.clipToOutline = true
         }
     }
 }
