@@ -1,7 +1,7 @@
 package com.greylabsdev.pexwalls.presentation.screen.photo
 
 import com.greylabsdev.pexwalls.presentation.model.PhotoModel
-import org.koin.android.viewmodel.dsl.viewModel
+import org.koin.core.module.dsl.viewModel
 import org.koin.dsl.module
 
 val photoModule = module {
