@@ -5,8 +5,8 @@ import android.content.Context
 import android.util.AttributeSet
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.greylabsdev.pexwalls.R
+import com.greylabsdev.pexwalls.databinding.ViewProgressBinding
 import com.greylabsdev.pexwalls.presentation.ext.addEndListener
-import kotlinx.android.synthetic.main.view_progress.view.*
 
 class ProgressView @JvmOverloads constructor(
     context: Context,
@@ -15,33 +15,35 @@ class ProgressView @JvmOverloads constructor(
 ) : ConstraintLayout(context, attrs, defStyleAttr) {
 
     private val animator: ViewAnimator
+    private val binding: ViewProgressBinding
 
     init {
         inflate(context, R.layout.view_progress, this)
+        binding = ViewProgressBinding.bind(this)
         animator = ViewAnimator()
         startLoadingAnimation()
     }
 
     private fun startLoadingAnimation() {
         val scaleFirst = AnimatorSet().apply {
-            play(animator.shiftUp(p_letter_tv, 450))
-                .with(animator.shiftDown(w_letter_tv, 450))
-                .with(animator.fadeIn(divider_v, 450))
+            play(animator.shiftUp(binding.pLetterTv, 450))
+                .with(animator.shiftDown(binding.wLetterTv, 450))
+                .with(animator.fadeIn(binding.dividerV, 450))
         }
         val scaleBack = AnimatorSet().apply {
-            play(animator.shiftBack(w_letter_tv, 450))
-                .with(animator.shiftBack(p_letter_tv, 450))
-                .with(animator.fadeOut(divider_v, 450))
+            play(animator.shiftBack(binding.pLetterTv, 450))
+                .with(animator.shiftBack(binding.wLetterTv, 450))
+                .with(animator.fadeOut(binding.dividerV, 450))
         }
         val scaleBackSecond = AnimatorSet().apply {
-            play(animator.shiftBack(p_letter_tv, 450))
-                .with(animator.shiftBack(w_letter_tv, 450))
-                .with(animator.fadeOut(divider_v, 450))
+            play(animator.shiftBack(binding.pLetterTv, 450))
+                .with(animator.shiftBack(binding.wLetterTv, 450))
+                .with(animator.fadeOut(binding.dividerV, 450))
         }
         val scaleSecond = AnimatorSet().apply {
-            play(animator.shiftUp(w_letter_tv, 450))
-                .with(animator.shiftDown(p_letter_tv, 450))
-                .with(animator.fadeIn(divider_v, 450))
+            play(animator.shiftUp(binding.pLetterTv, 450))
+                .with(animator.shiftDown(binding.wLetterTv, 450))
+                .with(animator.fadeIn(binding.dividerV, 450))
         }
         val scaleFull = AnimatorSet().apply {
             play(scaleFirst).before(scaleBack)
